requirements.txt

openai>=1.0.0
requests
playwright
playwright-stealth
pdfplumber
ddddocr
Pillow
